Raw Data/requirements.txt
tweepy
PyYAML

